ownerships and permissions added and fixed

+++ services/service_views/remove_feature_from_service.py
@@ -1,6 +1,8 @@
 from rest_framework import status
 from rest_framework.response import Response
-from services.service_models.feature import ServiceFeature
+from services.service_models.feature import ServiceFeature, Feature
+from services.service_views.add_feature_to_service import validate_ownership
+
 
 def remove_feature(self, request, pk=None):
     service = self.get_object()
@@ -11,6 +13,10 @@ def remove_feature(self, request, pk=None):
 
     try:
         # Ensure the feature is linked to the service
+        feature = Feature.objects.get(pk=feature_id)
+        service_feature = ServiceFeature.objects.filter(service=service, feature=feature).first()
+        validate_ownership(user=request.user, service=service, feature=feature, service_feature=service_feature)
+
         service_feature = ServiceFeature.objects.get(service=service, feature_id=feature_id)
         service_feature.delete()  # Delete the relationship
         return Response({"message": "Feature removed from service successfully."}, status=status.HTTP_200_OK)
